CRUD/requirements.txt
asgiref==3.4.1
dj-rest-auth==2.2.2
Django==4.0.1
django-crispy-forms==1.13.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
